+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Added affiliateCampaignId, affiliateReferenceId and contextualLocation configuration options.
 * Added Bulk Migrate Listings operation to the Inventory service.
 * Support Metadata API version 1.1.0.
+* Added Get Shipping Rate Tables operation to the Account service. Note that eBay have not officaly released this
+  operation so it may be subject to change.
 
 ### Breaking changes